Add multi-file synchronization
Wow, this marks a **big** milestone in bifimu history. Right now, I implemented multi-file synchronization -- meaning that instead of only synchronizing the contents of one file, bifimu will be able to synchronize a variable number of files. Sadly, in order to get this to work, I had to get rid of neighborhoodhacker's argument parser.